Divide program to modules
